% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filter_H}
\alias{filter_H}
\title{This function is for filtering data}
\usage{
filter_H(data, columns)
}
\arguments{
\item{data}{must be a data set that contains character data. Hormone data should be organized in each column.}

\item{columns}{include the hormones that you will be selecting for the function. You can include the hormones that you want to filtering.}
}
\value{
filter columns with the concentration of VTG and E2 (numeric)
}
\description{
Use for filter the hormones concentration that is on the columns in the data set.
Returns hormone concentration of the filter data request.
}
