% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{function_DR}
\alias{function_DR}
\title{This function is for select specific data}
\usage{
function_DR(data, columns)
}
\arguments{
\item{data}{must be a data set that contains numeric values in some columns.}

\item{columns}{include the hormones that you will be selecting for the function}
}
\value{
columns with the concentration of the hormones (numeric)
}
\description{
Use for select the hormone concentration columns that are in data set.
Returns concentration values of the selected hormones.
}
