% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_model.R
\name{model_ln}
\alias{model_ln}
\title{this function is for statistical analysis by using a linear model.}
\usage{
model_ln(data, column_1, column_2)
}
\arguments{
\item{data}{must be a data set that contains numeric data.}

\item{column_1}{must include numeric data}

\item{column_2}{must include numeric data}
}
\value{
the linear model with the statistical analysis.
}
\description{
Use for do a linear model in the hormone data set.
Returns a linear model with the predictor and the response variables.
}
