% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bar_plot.R
\name{Plot}
\alias{Plot}
\title{Use for plot a graph of the hormones concentration that are in the data set.
Returns plot of the hormone concentration of the data selected.}
\usage{
Plot(data, column_1, column_2)
}
\arguments{
\item{data}{must be a data set that contains numeric data for plot the columns.}

\item{column_1}{include here the data that you want to use for the x aes, this should be numerical. In this case you will be using the hormone concentrations.}

\item{column_2}{include here the data that you want to use for the y aes, this should be a character data. In this case you will be using the Note.}
}
\value{
a bar plot with the data that you selected
}
\description{
Use for plot a graph of the hormones concentration that are in the data set.
Returns plot of the hormone concentration of the data selected.
}
