% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_two.R
\name{Plot_two}
\alias{Plot_two}
\title{This function is for make a graph by using two variables.}
\usage{
Plot_two(data, column_1, column_2, column_3)
}
\arguments{
\item{data}{must be a data set that contains numeric data for plot the aes.}

\item{column_1}{include here the data that you want to use for the x aes. In this case you will be using the hormone concentrations.}

\item{column_2}{include here the data that you want to use for the y aes. In this case you will be using the hormone concentrations.}

\item{column_3}{include here character data for color. In this case you will be using Note.}
}
\value{
the graph with the data that that you selected (numeric).
}
\description{
Use for plot hormones concentration that are in a the data set.
Returns plot of the hormone concentration of the data selected.
}
