% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filter_H}
\alias{filter_H}
\title{This function is for filtering data}
\usage{
filter_H(data, columns)
}
\arguments{
\item{data}{in the date is expecting that each hormone is in each column}

\item{columns}{include the hormones that you will be selecting for the function}
}
\value{
filter columns with the concentration of VTG and E2 (numeric)
}
\description{
Use for filter the hormones concentration that is on the columns in the data set.
Returns hormone concentration of the filter data request.
}
